Add Day 3
The parser for Day 3 was simpler than the parser in Day 2, but it
required a bit of thought on how to retain enough information about the
codes, while retaining the grid size and not polluting it.

Codes are compressed into a single cell, with remaining cells in their
digit length being marked empty to retain proper grid size. However
codes also carry a length property ('digits') to allow for later
boundary checks to be done properly for identifying adjacent symbols.

The second part of the puzzle is completed by accumulating each adjacent
symbol in a map keyed on the symbol position. Paired codes accumulate in
pairs, and are filtered out and processed after iterating over the
schematic.