Add day 5
The second part of this day was a brain teaser. It is possible to brute
force it, but the ranges are so large that it can blow a stack or take a
very very long time. Instead I implemented a solution that looks for
intersects and differences between the ranges to identify and translate
seeds with destinations and seeds without.